Separated component and view
