--- src/components/TopNavbarLogin.vue
@@ -1,11 +1,0 @@
-<template>
-   <div class="top-navbar bg-white drop-shadow-sm px-3 py-4">
-      <img src="../assets/logo.png" alt="image" class="mx-auto" />
-    </div>
-  
-</template>
-
-<script >
- 
- 
-</script>
